refactoring api-product import list logic
