+++ import-export-cli/cmd/apiProducts.go
@@ -19,11 +19,9 @@
 package cmd
 
 import (
-	"encoding/json"
-	"errors"
 	"fmt"
+	"github.com/wso2/product-apim-tooling/import-export-cli/impl"
 	"io"
-	"net/http"
 	"os"
 	"text/template"
 
@@ -136,57 +134,13 @@ func executeApiProductsCmd(credential credentials.Credential) {
 	}
 
 	// Unified Search endpoint from the config file to search API Products
-	unifiedSearchEndpoint := utils.GetUnifiedSearchEndpointOfEnv(listApiProductsCmdEnvironment, utils.MainConfigFilePath)
-	_, apiProducts, err := GetAPIProductList(listApiProductsCmdQuery, listApiProductsCmdLimit, accessToken, unifiedSearchEndpoint)
+	_, apiProducts, err := impl.GetAPIProductList(accessToken, listApiProductsCmdEnvironment, listApiProductsCmdQuery,
+		listApiProductsCmdLimit)
 	if err == nil {
 		printAPIProducts(apiProducts, listApiProductsCmdFormat)
 	} else {
 		utils.Logln(utils.LogPrefixError+"Getting List of API Products", err)
 	}
-}
-
-// GetAPIProductList
-// @param query : String to be matched against the API Product names
-// @param accessToken : Access Token for the environment
-// @param unifiedSearchEndpoint : Unified Search Endpoint for the environment to retreive API Product list
-// @return count (no. of API Products)
-// @return array of API Product objects
-// @return error
-func GetAPIProductList(query, limit, accessToken, unifiedSearchEndpoint string) (count int32, apiProducts []utils.APIProduct, err error) {
-	headers := make(map[string]string)
-	headers[utils.HeaderAuthorization] = utils.HeaderValueAuthBearerPrefix + " " + accessToken
-
-	// To filter API Products from unified search
-	unifiedSearchEndpoint += "?query=type:\"" + utils.DefaultApiProductType + "\""
-
-	// Setting up the query parameter and limit parameter
-	if query != "" {
-		unifiedSearchEndpoint += " " + query
-	}
-	if limit != "" {
-		unifiedSearchEndpoint += "&limit=" + limit
-	}
-	utils.Logln(utils.LogPrefixInfo+"URL:", unifiedSearchEndpoint)
-	resp, err := utils.InvokeGETRequest(unifiedSearchEndpoint, headers)
-
-	if err != nil {
-		utils.HandleErrorAndExit("Unable to connect to "+unifiedSearchEndpoint, err)
-	}
-
-	utils.Logln(utils.LogPrefixInfo+"Response:", resp.Status())
-
-	if resp.StatusCode() == http.StatusOK {
-		apiProductListResponse := &utils.APIProductListResponse{}
-		unmarshalError := json.Unmarshal([]byte(resp.Body()), &apiProductListResponse)
-
-		if unmarshalError != nil {
-			utils.HandleErrorAndExit(utils.LogPrefixError+"invalid JSON response", unmarshalError)
-		}
-		return apiProductListResponse.Count, apiProductListResponse.List, nil
-	} else {
-		return 0, nil, errors.New(string(resp.Body()))
-	}
-
 }
 
 // printAPIProducts
